Resolved formatting issues
- Resolved formatting issues
- Added cache.go and model.go files
